Further improvements to the program
Created a new python file gui.py which will run the graphics version of the program. This does not eradicate the shell version however because both will coincide and the user will choose which version to run. 

main.py is renamed to mainShell.py. Might rename gui.py (or create mainGUI.py). 

There are more prompts that will allow for create/deleting/reading schedules from and to the schedule.txt file.

ioControl.py has more power over reading and writing to file which works with prompts.py

gui.py uses tkinter and tkcalendar to operate. Still a work in progress because using tutorials/documentation to learn it.